requirements.txt
streamlit
openai
langchain
pinecone
pypdf
python-dotenv
tiktoken
